move skeleton files to skeleton folder
