+++ cocos2d/core/3d/index.js
@@ -6,12 +6,12 @@ if (!CC_TEST && (!CC_EDITOR || !Editor.isMainProcess)) {
 }
 
 require('./CCModel');
-require('./CCSkeleton');
-require('./CCSkeletonAnimationClip');
+require('./skeleton/CCSkeleton');
+require('./skeleton/CCSkeletonAnimationClip');
 require('./actions');
 
 if (!CC_EDITOR || !Editor.isMainProcess) {
-    require('./CCSkeletonAnimation');
-    require('./CCSkinnedMeshRenderer');
+    require('./skeleton/CCSkeletonAnimation');
+    require('./skeleton/CCSkinnedMeshRenderer');
     require('./CCLightComponent');
 }
