+++ cocos2d/core/3d/skeleton/CCSkeleton.js
